Fix link for second additional resource
The link for the second additional resource, "Learn the new Firefox Dev Tools in this [Video](https://youtu.be/yznVkCuohGg) by Wes Bos" leads to a video called "CSS GRID: Dev Tools — 4 of 25", which does not seem relevant and while it uses Firefox Dev Tools, it seems very specified and not helpful for the point we are at in the course (and we've never even gone over css grid). I looked for a better video by Wes Bos on the topic but couldn't find one; my suggestion would be another YouTube video altogether actually about "new Firefox Dev Tools".